+++ src/resources/HierarchyData.cpp
@@ -43,9 +43,9 @@ void HierarchyData::loadFromJSON(const json &jsonData) {
 
   // Children found, load them
   auto jsonChildren = jsonData["children"];
-  int size = jsonChildren.size();
+  size_t size = jsonChildren.size();
   children.resize(size);
-  for (int i = 0; i < size; i++) {
+  for (size_t i = 0; i < size; i++) {
     children[i] = std::make_shared<HierarchyData>();
     children[i]->loadFromJSON(jsonChildren.at(i));
   }
